requirements.txt
jupyter
transformers
accelerate
pillow
google-cloud-documentai
matplotlib
openai
python-dotenv
nbstripout
pre-commit
requests


